requirements.txt
arrow==1.0.3
asgiref==3.2.10
automium==0.2.6
automium-web==0.1.1
binaryornot==0.4.4
boto3==1.16.34
botocore==1.19.34
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cookiecutter==1.7.2
Django==3.1.4
django-appconf==1.0.4
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-imagekit==4.0.2
django-js-asset==1.2.2
django-storages==1.10.1
Flask==1.1.2
gTTS==2.2.2
gunicorn==20.1.0
idna==2.10
image==1.5.33
itsdangerous==1.1.0
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
legacy==0.1.6
MarkupSafe==1.1.1
numpy==1.20.2
pilkit==2.0
Pillow==8.0.1
playsound==1.2.2
poyo==0.5.0
psycopg2==2.8.6
python-dateutil==2.8.1
python-slugify==4.0.1
pytz==2020.1
requests==2.25.1
s3transfer==0.3.3
six==1.15.0
sqlparse==0.3.1
text-unidecode==1.3
urllib3==1.26.2
Werkzeug==1.0.1
whitenoise==5.2.0
wxPython==4.1.1